Small cleanup, moving stuff to utils
